ui/requirements.txt
Flask==2.2.3
requests==2.28.2
PyYAML==6.0
python-openstackclient==6.0.0
python-novaclient==18.1.0
python-glanceclient==4.0.0
python-neutronclient==8.1.0
python-keystoneclient==5.0.1
gunicorn==20.1.0
Flask-WTF==1.1.1
python-dateutil==2.8.2
